refactor(services): convert comment service to pure factory function
